Only place defender if there is no existing defender on that square.
Note that tutorial solved this differently by placing a box collider on the defenders that prevented a click from getting through to the Core Game Area.

I stuck with my solution because a) I don't think we'll have enough game defenders that looping through and comparing positionns would become a performance issue and b) I think it's a little more fool proof than putting box colliders on defenders. For b), having a box collider on each defender means that there is one more thing to remember when creating new defenders and also the possibility that the box collider isn't quite the right size, or if we change our grid size on the core game object we have to go through each defender to change the collider box size.